backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Week 2: Satellite Imagery
sentinelhub>=3.10.0
rasterio>=1.3.10
numpy>=1.26.0
pillow>=10.0.0
python-multipart>=0.0.9

# Week 3: Image Pre-processing
opencv-python>=4.8.0
scikit-image>=0.22.0
scipy>=1.11.0
